Changing float to int

+++ validate_quantity.py
@@ -1,7 +1,7 @@
 def validate_quantity(value, column_name):
 
     try:
-        quantity = float(value.get(column_name))
+        quantity = int(value.get(column_name))
         if quantity <= 0:
             return False 
         else:
